move postMessageForResponse to puc
