+++ packages/wallet-sdk/src/core/communicator/CrossDomainCommunicator.ts
@@ -1,4 +1,4 @@
-import { Message, MessageID } from '../message';
+import { Message } from '../message';
 import { standardErrors } from ':core/error';
 
 export abstract class CrossDomainCommunicator {
@@ -18,7 +18,6 @@ export abstract class CrossDomainCommunicator {
   protected disconnect(): void {
     this.connected = false;
     window.removeEventListener('message', this.eventListener.bind(this));
-    this.rejectWaitingRequests();
   }
 
   protected peerWindow: Window | null = null;
@@ -38,42 +37,8 @@ export abstract class CrossDomainCommunicator {
     this.peerWindow.postMessage(message, targetOrigin);
   }
 
-  async postMessageForResponse(message: Message): Promise<Message> {
-    this.postMessage(message);
-    return new Promise((resolve, reject) => {
-      this.requestMap.set(message.id, {
-        resolve,
-        reject,
-      });
-    });
-  }
-
-  private requestMap = new Map<
-    MessageID,
-    {
-      resolve: (_: Message) => void;
-      reject: (_: Error) => void;
-    }
-  >();
-
   private eventListener(event: MessageEvent<Message>) {
     if (event.origin !== this.url?.origin) return;
-
-    const message = event.data;
-    const { requestId } = message;
-    if (!requestId) {
-      this.handleIncomingEvent(event);
-      return;
-    }
-
-    this.requestMap.get(requestId)?.resolve?.(message);
-    this.requestMap.delete(requestId);
-  }
-
-  private rejectWaitingRequests() {
-    this.requestMap.forEach(({ reject }) => {
-      reject(standardErrors.provider.userRejectedRequest('Request rejected'));
-    });
-    this.requestMap.clear();
+    this.handleIncomingEvent(event);
   }
 }
